+++ libsrc/libbioimg/formats_api/bim_buffer.h
@@ -14,25 +14,27 @@
 #ifndef BIM_BUFFER_H
 #define BIM_BUFFER_H
 
+#include <stddef.h>
+
 // simple buffer with storage entity byte, allocates and deallocated memory space
 class DimBuffer {
 public:
-  DimBuffer( unsigned int size=0 );
-  DimBuffer( unsigned int size, unsigned char fill_byte );
+  DimBuffer( size_t size=0 );
+  DimBuffer( size_t size, unsigned char fill_byte );
   ~DimBuffer();
 
-  virtual bool allocate( unsigned int size );
+  virtual bool allocate( size_t size );
   virtual void fill( unsigned char fill_byte = 0 );
   void free( );
 
-  inline unsigned int size() const { return buf_size; }
+  inline size_t size() const { return buf_size; }
   inline unsigned char *buffer() const { return buf; }
   inline unsigned char *bytes() const { return buf; }
-  inline unsigned char &operator[](int i) const { return *(buf + i); }
+  inline unsigned char &operator[](ptrdiff_t i) const { return *(buf + i); }
 
 private:
   unsigned char *buf;
-  unsigned int buf_size;
+  size_t buf_size;
 
 };
 
@@ -42,30 +44,30 @@ template <typename T>
 class DTypedBuffer : public DimBuffer {
 public:
   
-  DTypedBuffer( unsigned int size=0 ): DimBuffer( sizeof(T)*size ) { }
+  DTypedBuffer( size_t size=0 ): DimBuffer( sizeof(T)*size ) { }
 
-  DTypedBuffer( unsigned int size, const T &fill_value ): DimBuffer( sizeof(T)*size ) {
+  DTypedBuffer( size_t size, const T &fill_value ): DimBuffer( sizeof(T)*size ) {
     fill( fill_value );
   }
 
   inline T* buffer() const { return (T*) bytes(); }
 
-  bool allocate( unsigned int size ) {
+  bool allocate( size_t size ) {
     return DimBuffer::allocate( sizeof(T)*size );
   }
   
   void fill( const T &fill_value ) {
     if (size()==0) return;
     T *p = (T*) bytes();
-    for (unsigned int i=0; i<size(); ++i) 
+    for (size_t i=0; i<size(); ++i) 
       p[i] = fill_value;
   }
 
-  inline unsigned int size() const { 
+  inline size_t size() const { 
     return ( DimBuffer::size()/sizeof(T) ); 
   }
 
-  inline T &operator[](int i) const { 
+  inline T &operator[](ptrdiff_t i) const { 
     T *p = (T*) bytes();
     return *(p + i); 
   }
